+++ fiboa_cli/parquet.py
@@ -8,6 +8,8 @@ from .types import get_geopandas_dtype, get_pyarrow_type_for_geopandas, get_pyar
 from .util import log, load_fiboa_schema, load_file, merge_schemas
 from .geopandas import to_parquet
 
+ROW_GROUP_SIZE = 25000
+
 def create_parquet(data, columns, collection, output_file, config, missing_schemas = {}, compression = None, geoparquet1 = False):
     # Load the data schema
     fiboa_schema = load_fiboa_schema(config)
@@ -84,6 +86,7 @@ def create_parquet(data, columns, collection, output_file, config, missing_schem
         coerce_timestamps = "ms",
         compression = compression,
         schema_version = "1.0.0" if geoparquet1 else "1.1.0",
+        row_group_size = ROW_GROUP_SIZE,
         write_covering_bbox = False if geoparquet1 else True
     )