+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Converter for Germany, Mecklenburg-Western Pomerania
 - Converter for Germany, Saarland
 - Converter for Germany, Saxony
+- Converter for Estonia (EuroCrops)
 
 ### Changed