+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v0.8.0] - 2024-11-12
+
 ### Added
 
 - Merge command: `fiboa merge`
@@ -329,7 +331,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - First release
 
-[Unreleased]: <https://github.com/fiboa/cli/compare/v0.7.0...main>
+[Unreleased]: <https://github.com/fiboa/cli/compare/v0.8.0...main>
+[v0.8.0]: <https://github.com/fiboa/cli/compare/v0.7.0...v0.8.0>
 [v0.7.0]: <https://github.com/fiboa/cli/compare/v0.6.0...v0.7.0>
 [v0.6.0]: <https://github.com/fiboa/cli/compare/v0.5.0...v0.6.0>
 [v0.5.0]: <https://github.com/fiboa/cli/compare/v0.4.0...v0.5.0>